requirements.txt
python-telegram-bot==13.15  # Version compatible avec Python 3.6
requests==2.28.2            # Version stable pour 3.6
beautifulsoup4==4.11.1      # Compatible 3.6
feedparser==6.0.10
python-dotenv==0.21.1       # Version pour 3.6
schedule==1.2.0
lxml==4.9.0
